+++ engine/flink/components/base/src/it/scala/pl/touk/nussknacker/engine/flink/util/transformer/UnionTransformerSpec.scala
@@ -3,6 +3,7 @@ package pl.touk.nussknacker.engine.flink.util.transformer
 import cats.data.NonEmptyList
 import com.typesafe.config.{ConfigFactory, ConfigValueFactory}
 import com.typesafe.scalalogging.LazyLogging
+import org.apache.flink.api.common.typeinfo.TypeInformation
 import org.scalatest._
 import pl.touk.nussknacker.engine.api.deployment.DeploymentData
 import pl.touk.nussknacker.engine.api.{MetaData, ProcessVersion, StreamMetaData}
@@ -22,6 +23,9 @@ import pl.touk.nussknacker.test.VeryPatientScalaFutures
 
 import java.util
 
+import scala.reflect.runtime.universe._
+import scala.reflect.ClassTag
+
 class UnionTransformerSpec extends FunSuite with BeforeAndAfterAll with Matchers with FlinkSpec with EitherValues with LazyLogging with VeryPatientScalaFutures {
 
   import org.apache.flink.streaming.api.scala._
@@ -35,8 +39,31 @@ class UnionTransformerSpec extends FunSuite with BeforeAndAfterAll with Matchers
 
   private val OutVariableName = "outVar"
 
-  test("should unify streams") {
-    val scenario =  EspProcess(MetaData("sample-union-memo", StreamMetaData()), NonEmptyList.of[SourceNode](
+  val data = List("10", "20", "30", "40")
+
+  test("should unify streams with union-memo") {
+    val scenario = EspProcess(MetaData("sample-union-memo", StreamMetaData()), NonEmptyList.of[SourceNode](
+      GraphBuilder.source("start-foo", "source")
+        .branchEnd(BranchFooId, UnionNodeId),
+      GraphBuilder.source("start-bar", "noopSource")
+        .branchEnd(BranchBarId, UnionNodeId),
+      GraphBuilder
+        .branch(UnionNodeId, "union-memo", Some(OutVariableName),
+          List(
+            BranchFooId -> List("key" -> "'fooKey'", "value" -> "#input"),
+            BranchBarId -> List("key" -> "'barKey'", "value" -> "#input")
+          ), "stateTimeout" -> "T(java.time.Duration).parse('PT1M')"
+        )
+        .processorEnd("end", "mockService", "all" -> s"#$OutVariableName.$BranchFooId")
+    ))
+
+    run(scenario, data)
+    MockService.data shouldBe data
+  }
+
+  test("should unify streams with union when one branch is empty") {
+    MockService.clear()
+    val scenario = EspProcess(MetaData("sample-union", StreamMetaData()), NonEmptyList.of[SourceNode](
       GraphBuilder.source("start-foo", "source")
         .branchEnd(BranchFooId, UnionNodeId),
       GraphBuilder.source("start-bar", "noopSource")
@@ -44,28 +71,85 @@ class UnionTransformerSpec extends FunSuite with BeforeAndAfterAll with Matchers
       GraphBuilder
         .branch(UnionNodeId, "union", Some(OutVariableName),
           List(
-            BranchFooId -> List(
-              "key" -> "'fooKey'",
-              "value" -> "#input"
-            ),
-            BranchBarId -> List(
-              "key" -> "'fooKey'",
-              "value" -> "#input"
-            )
-          )
+            BranchFooId -> List("Output expression" -> "{a: #input}"),
+            BranchBarId -> List("Output expression" -> "{a: '123'}"))
         )
-        .processorEnd("end", "mockService", "all" -> s"#$OutVariableName.$BranchFooId")
+        .processorEnd("end", "mockService", "all" -> s"#$OutVariableName.a")
     ))
 
-    val data = List("10")
     run(scenario, data)
     MockService.data shouldBe data
   }
 
-  def run(process: EspProcess, data: List[String]): Unit = {
+  test("should unify streams with union when both branches emit data") {
+    MockService.clear()
+    val scenario = EspProcess(MetaData("sample-union", StreamMetaData()), NonEmptyList.of[SourceNode](
+      GraphBuilder.source("start-foo", "source")
+        .branchEnd(BranchFooId, UnionNodeId),
+      GraphBuilder.source("start-bar", "source")
+        .branchEnd(BranchBarId, UnionNodeId),
+      GraphBuilder
+        .branch(UnionNodeId, "union", Some(OutVariableName),
+          List(
+            BranchFooId -> List("Output expression" -> "{a: #input}"),
+            BranchBarId -> List("Output expression" -> "{a: '123'}"))
+        )
+        .processorEnd("end", "mockService", "all" -> s"#$OutVariableName.a")
+    ))
+
+    run(scenario, data)
+    MockService.data.size shouldBe data.size * 2
+    MockService.data.toSet shouldBe data.toSet + "123"
+  }
+
+  test("should throw when contexts are different") {
+    MockService.clear()
+    val scenario = EspProcess(MetaData("sample-union", StreamMetaData()), NonEmptyList.of[SourceNode](
+      GraphBuilder.source("start-foo", "source")
+        .branchEnd(BranchFooId, UnionNodeId),
+      GraphBuilder.source("start-bar", "noopSource")
+        .branchEnd(BranchBarId, UnionNodeId),
+      GraphBuilder
+        .branch(UnionNodeId, "union", Some(OutVariableName),
+          List(
+            BranchFooId -> List("Output expression" -> "{a: #input}"),
+            BranchBarId -> List("Output expression" -> "{b: 123}")
+          )
+        )
+        .processorEnd("end", "mockService", "all" -> s"#$OutVariableName.a")
+    ))
+
+    intercept[IllegalArgumentException] {
+      run(scenario, data)
+    }.getMessage should include("All branch values must be of the same")
+  }
+
+  test("should not throw when one branch emits error") {
+    MockService.clear()
+    val scenario = EspProcess(MetaData("sample-union", StreamMetaData()), NonEmptyList.of[SourceNode](
+      GraphBuilder.source("start-foo", "source")
+        .branchEnd(BranchFooId, UnionNodeId),
+      GraphBuilder.source("start-bar", "source")
+        .branchEnd(BranchBarId, UnionNodeId),
+      GraphBuilder
+        .branch(UnionNodeId, "union", Some(OutVariableName),
+          List(
+            BranchFooId -> List("Output expression" -> "#input"),
+            BranchBarId -> List("Output expression" -> "#input / (#input % 4)")
+          )
+        )
+        .processorEnd("end", "mockService", "all" -> s"#$OutVariableName")
+    ))
+
+    run(scenario, List(10, 20, 30, 40))
+    MockService.data.size shouldBe 6
+    MockService.data.toSet shouldBe Set(5, 10, 15, 20, 30, 40)
+  }
+
+  def run[T: ClassTag : TypeTag : TypeInformation](process: EspProcess, data: List[T]): Unit = {
     val env = flinkMiniCluster.createExecutionEnvironment()
     val finalConfig = ConfigFactory.load().withValue("components.base", ConfigValueFactory.fromMap(new util.HashMap[String, AnyRef]()))
-    val resolvedConfig =  new DefaultModelConfigLoader().resolveInputConfigDuringExecution(finalConfig, getClass.getClassLoader).config
+    val resolvedConfig = new DefaultModelConfigLoader().resolveInputConfigDuringExecution(finalConfig, getClass.getClassLoader).config
     val modelData = LocalModelData(resolvedConfig, new BaseSampleConfigCreator(data))
     val registrar = FlinkProcessRegistrar(new FlinkProcessCompiler(modelData), ExecutionConfigPreparer.unOptimizedChain(modelData))
     registrar.register(new StreamExecutionEnvironment(env), process, ProcessVersion.empty, DeploymentData.empty)